+++ apps/website/package.json
@@ -9,11 +9,14 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@headlessui/react": "1.7.17",
+    "@heroicons/react": "2.0.18",
     "@namehash/nameguard": "workspace:*",
     "@namehash/nameguard-react": "workspace:*",
     "classcat": "5.0.4",
     "next": "latest",
     "react": "latest",
+    "react-debounce-input": "3.3.0",
     "react-dom": "latest"
   },
   "devDependencies": {
