+++ apps/website/src/app/report/[slug]/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <p>Loading</p>;
+}
