+++ packages/nameguard-react/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@namehash/nameguard-react",
+  "version": "0.0.2",
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch --clean=false"
@@ -21,7 +22,7 @@
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "dependencies": {
-    "@namehash/nameguard": "workspace:*",
+    "@namehash/nameguard": "0.0.5",
     "classcat": "5.0.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
