+++ packages/nameguard-react/src/index.ts
@@ -1,3 +1,4 @@
 export { default as Report } from "./Report";
 export { default as Summary } from "./Summary";
 export { default as Shield } from "./Shield";
+export { default as Status } from "./Status";
